Add Length related tests
