+++ codyze-backends/cpg/src/test/kotlin/de/fraunhofer/aisec/codyze/backends/cpg/coko/evaluators/NeverEvaluationTest.kt
@@ -13,9 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.fraunhofer.aisec.codyze.backends.cpg
+package de.fraunhofer.aisec.codyze.backends.cpg.coko.evaluators
 
 import de.fraunhofer.aisec.codyze.backends.cpg.coko.CokoCpgBackend
+import de.fraunhofer.aisec.codyze.backends.cpg.createCpgConfiguration
+import de.fraunhofer.aisec.codyze.backends.cpg.dummyRule
 import de.fraunhofer.aisec.codyze.specificationLanguages.coko.core.EvaluationContext
 import de.fraunhofer.aisec.codyze.specificationLanguages.coko.core.Finding
 import de.fraunhofer.aisec.codyze.specificationLanguages.coko.core.dsl.definition
@@ -30,24 +32,28 @@ import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 import kotlin.test.assertTrue
 
-class OnlyEvaluationTest {
+class NeverEvaluationTest {
 
     class FooModel {
         fun first(i: Any) = op {
-            definition("Foo.fun") {
+            definition("Foo.first") {
                 signature(i)
             }
         }
     }
 
     @Test
-    fun `test simple only`() {
+    fun `test never with violation`() {
         val fooInstance = FooModel()
 
-        val backend = CokoCpgBackend(config = createCpgConfiguration(testFile))
+        val backend = CokoCpgBackend(config = createCpgConfiguration(violationFile))
 
         with(backend) {
-            val evaluator = only(fooInstance.first(0..10))
+            // Evaluator does not allow calls to `first` with -1 or a number between 1230 and 1240
+            val evaluator = never(
+                fooInstance.first(-1),
+                fooInstance.first(1230..1240)
+            )
             val findings = evaluator.evaluate(
                 EvaluationContext(
                     rule = ::dummyRule,
@@ -62,18 +68,52 @@ class OnlyEvaluationTest {
         }
     }
 
+    @Test
+    fun `test never with no violations`() {
+        val fooInstance = FooModel()
+
+        val backend = CokoCpgBackend(config = createCpgConfiguration(passFile))
+
+        with(backend) {
+            // Evaluator does not allow calls to `first` with -1 or a number between 1230 and 1240
+            val evaluator = never(
+                fooInstance.first(-1),
+                fooInstance.first(1230..1240)
+            )
+            val findings = evaluator.evaluate(
+                EvaluationContext(
+                    rule = ::dummyRule,
+                    parameterMap = ::dummyRule.valueParameters.associateWith { fooInstance }
+                )
+            )
+
+            assertTrue("There were no findings which is unexpected") { findings.isNotEmpty() }
+
+            assertTrue("Not all findings are passes which is unexpected: ${findings.joinToString()}") {
+                findings.all { it.kind == Finding.Kind.Pass }
+            }
+
+            assertEquals(1, findings.size, "Found ${findings.size} finding(s) instead of one pass finding")
+        }
+    }
+
     companion object {
 
-        lateinit var testFile: Path
+        lateinit var violationFile: Path
+        lateinit var passFile: Path
 
         @BeforeAll
         @JvmStatic
         fun startup() {
             val classLoader = OnlyEvaluationTest::class.java.classLoader
 
-            val testFileResource = classLoader.getResource("OnlyEvaluationTest/SimpleOnly.java")
-            assertNotNull(testFileResource)
-            testFile = testFileResource.toURI().toPath()
+            val violationFileResource = classLoader.getResource("NeverEvaluationTest/NeverViolation.java")
+            assertNotNull(violationFileResource)
+            violationFile = violationFileResource.toURI().toPath()
+
+            val passFileResource = classLoader.getResource("NeverEvaluationTest/NeverPass.java")
+            assertNotNull(passFileResource)
+            passFile = passFileResource.toURI().toPath()
         }
     }
 }
