+++ codyze-backends/cpg/src/test/kotlin/de/fraunhofer/aisec/codyze/backends/cpg/coko/ordering/NfaDfaConstructionTest.kt
@@ -13,10 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.fraunhofer.aisec.codyze.backends.cpg
+package de.fraunhofer.aisec.codyze.backends.cpg.coko.ordering
 
 import de.fraunhofer.aisec.codyze.backends.cpg.coko.CokoCpgBackend
-import de.fraunhofer.aisec.codyze.backends.cpg.coko.ordering.toNfa
 import de.fraunhofer.aisec.codyze.specificationLanguages.coko.core.dsl.*
 import de.fraunhofer.aisec.codyze.specificationLanguages.coko.core.ordering.*
 import de.fraunhofer.aisec.cpg.analysis.fsm.DFA
@@ -48,7 +47,7 @@ class NfaDfaConstructionTest {
     }
 
     private val baseName =
-        "de.fraunhofer.aisec.codyze.backends.cpg.NfaDfaConstructionTest\$TestClass"
+        "de.fraunhofer.aisec.codyze.backends.cpg.coko.ordering.NfaDfaConstructionTest\$TestClass"
 
     private fun orderExpressionToNfa(block: Order.() -> Unit): NFA {
         val order = Order().apply(block)
